HashCode2014

Documentation for HashCode2014.

HashCode2014.HashCode2014

HashCode2014.City

HashCode2014.Junction

HashCode2014.Street

HashCode2014.read_city

HashCode2014.read_solution

HashCode2014.write_city

HashCode2014.write_solution

HashCode2014.HashCode2014 — Module

HashCode2014

Lightweight package designed to store the data of the 2014 Google Hash Code.

Challenge description: https://storage.googleapis.com/coding-competitions.appspot.com/HC/2014/hashcode2014_final_task.pdf

Challenge data: https://storage.googleapis.com/coding-competitions.appspot.com/HC/2014/paris_54000.txt

HashCode2014.City — Type

City

Store a city made of Junctions and Streets, along with additional instance parameters.

Fields

N::Int: number of junctions

M::Int: number of streets

T::Int: time allotted for the car itineraries (in seconds)

C::Int: number of cars in the fleet

S::Int: junction at which all the cars are located initially

junctions::Vector{Junction}: list of junctions

streets::Vector{Street}: list of streets

HashCode2014.Junction — Type

Junction

Store a city junction.

Fields

i::Int: junction index

lat::Float64: latitude

long::Float64: longitude

HashCode2014.Street — Type

Street

Store an edge between two Junctions.

Fields

j::Int: street index

A::Int: index of the first junction

B::Int: index of the second junction

D::Int: direction, 1 for one-directional (A -> B only) and 2 for bi-directional

C::Int: time cost of traversing the street (in seconds)

L::Int: length of the street (in meters)

HashCode2014.read_city — Function

read_city(path)

Read and parse a City from a file located at path.

The default path is an artifact containing the official challenge data.

HashCode2014.read_solution — Method

read_solution(solution, path)

Read and parse a Solution from a file located at path.

HashCode2014.write_city — Method

write_city(city, path)

Write a City to a file located at path.

HashCode2014.write_solution — Method

write_solution(solution, path)

Write a Solution to a file located at path.